### What I did

This should be fast as ever.
This gets it a bit faster.

Before:

```
In [1]: %time from ape.exceptions import ContractLogicError
CPU times: user 236 ms, sys: 15.9 ms, total: 252 ms
Wall time: 251 ms
```

After:

```
In [1]: %time from ape.exceptions import ContractLogicError
CPU times: user 76.1 ms, sys: 13.5 ms, total: 89.7 ms
Wall time: 88.6 ms
```

It will be even better if I can improve `ape.logging`

### How I did it

Instead of importing from `eth_utils` (which is very slow), do simple logic myself.

### How to verify it

tests still pass should suffice

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated